+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Trades/TradeHistoryVM.cs
@@ -1,0 +1,6 @@
+namespace TD.WPF.ViewModels.Trades
+{
+    internal class TradeHistoryVM : AbstractDiaryVM
+    {
+    }
+}